Improvements all around
* Increase default angle to 12, which seems better for me
* Changes to better hold the trackpad into place
* Lower/rear walls now connect nicely
* Lower edge is slightly thinner

Notes:  I'd like to remove most of the lower wall (it's not necessary for holding the trackpad into place)
but if I do so the "arm" becomes too flimsy and bends too much to make it useful. I've left a module in
place in case others would like to experiment with it.